optional spawn
