+++ package.json
@@ -11,7 +11,6 @@
   "license": "MIT",
   "dependencies": {
     "any-shell-escape": "^0.1.1",
-    "duplexify": "^3.5.0",
     "dzen2-bin": "^1.0.2",
     "through2": "^2.0.3"
   },
